% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gheckman.R
\name{gheckman}
\alias{gheckman}
\title{Use this function to estimates Multivariate Switch model}
\usage{
gheckman(data, outcome, selection1 = NULL, selection2 = NULL,
  selection3 = NULL, selection4 = NULL, selection5 = NULL, group = NULL,
  zo3 = NULL, ShowInfo = TRUE, onlyTwostep = FALSE,
  opts = list(algorithm = "NLOPT_LD_TNEWTON", xtol_rel = 1e-08, print_level =
  1, maxeval = 1e+06), x0 = NULL)
}
\arguments{
\item{data}{frame containing the variables in the model}
}
\description{
Use this function to estimates Multivariate Switch model
}
